requirements.txt
# Core dependencies
praw==7.7.1
pandas==2.0.3
textblob==0.17.1
flask==2.3.3
requests==2.31.0

# Additional NLP libraries
vaderSentiment==3.3.2
nltk==3.8.1

# Data processing and storage
numpy==1.24.3

# API and web
flask-cors==4.0.0
python-dotenv==1.0.0

# Visualization (optional)
plotly==5.15.0
matplotlib==3.7.2
seaborn==0.12.2

# Testing
pytest==7.4.0
pytest-flask==1.2.0

# Scheduling and automation
schedule==1.2.0
APScheduler==3.10.4

# Development tools
black==23.7.0
flake8==6.0.0
